chore(team): split mailbox store inbox paths
